Move desktop log service to workbech/services
